refactor: move tests
* Creating a test class for the new PasscodeEndpoint class by extracting existing test methods.
* The git history for these tests is in the server/src/main/java/org/cloudfoundry/identity/uaa/login/LoginInfoEndpoint.java class.

[#186822654]
